Delete LargeSystemCache tweak according to SG TCPOptimizer.
As Microsoft has clearly pointed out “LargeSystemCache” is intended for systems used primarily as servers. For Workstation use it will usually be bad. The setting tells Windows that file caching has priority over memory for application use. For servers this makes sense, but for a workstation it can kill application performance.